separate files for attching texture/renderbuffer.
